+++ packages/wmr/test/fixtures/hmr/home.js
@@ -1,5 +1,18 @@
+import { FOO } from './store.js';
+import useCounter from './useCounter.js';
+
 function Home() {
-	return <p class="home">Home</p>;
+	const [count, increment] = useCounter();
+	return (
+		<div>
+			<p id="home-foo">{FOO}</p>
+			<p class="home">Home</p>
+			<p class="count">{count}</p>
+			<p class="increment" onClick={increment}>
+				Increment
+			</p>
+		</div>
+	);
 }
 
 export default Home;