+++ packages/wmr/test/fixtures/hmr/useCounter.js
@@ -1,0 +1,6 @@
+import { useState } from 'preact/hooks';
+
+export default function useCounter() {
+	const [count, setCount] = useState(0);
+	return [count, () => setCount(count + 1)];
+}