+++ packages/directory-plugin/CHANGELOG.md
@@ -1,8 +1,13 @@
 # @wmr-plugins/directory-import
 
+## 0.1.1
+
+### Patch Changes
+
+- [`3587239`](https://github.com/preactjs/wmr/commit/3587239c23c84f63a2f89d908553801b98a3b05a) [#380](https://github.com/preactjs/wmr/pull/380) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Add export maps and point at published files
+
 ## 0.1.0
+
 ### Minor Changes
-
-
 
 - [`bf6e97a`](https://github.com/preactjs/wmr/commit/bf6e97a5cf08876f08d3e11a123c53c8f34fdd20) [#344](https://github.com/preactjs/wmr/pull/344) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Initial release