+++ packages/sw-plugin/CHANGELOG.md
@@ -1,8 +1,13 @@
 # @wmr-plugins/service-worker
 
+## 0.1.1
+
+### Patch Changes
+
+- [`3587239`](https://github.com/preactjs/wmr/commit/3587239c23c84f63a2f89d908553801b98a3b05a) [#380](https://github.com/preactjs/wmr/pull/380) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Add export maps and point at published files
+
 ## 0.1.0
+
 ### Minor Changes
-
-
 
 - [`4bc66de`](https://github.com/preactjs/wmr/commit/4bc66de27c763a9fbf77b2d1a21b5362f810b087) [#346](https://github.com/preactjs/wmr/pull/346) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Initial release